# src/main/resources/static/boat4you_ws_entities_crud.openapi.yaml
openapi: 3.0.3
info:
  title: Boat4You Web Service
  description: |-
    API definition for Boat4You Web Service. \
    Contains operations related to Users
  version: 0.0.1
servers:
  - url: https://localhost:8443
    description: Local deployment
...
paths:
...
  /locations:
    get:
      tags:
        - Locations
      summary: Fetch all locations or search them by name
      description: Fetch all locations or search them by name
      operationId: getAllLocationsViews
      parameters:
        - in: query
          name: name
          schema:
            type: string
          required: false
          example: Split
          description: Name or part of name of the wanted location
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                description: Fetch all locations result
                required:
                  - content
                  - page
                properties:
                  content:
                    description: Location list
                    type: array
                    items:
                      $ref: '#/components/schemas/LocationViewDto'
                  page:
                    $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/InternalPageModel'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '500':
          description: Internal server error
  /countries:
    get:
      tags:
        - Locations
      summary: Fetch all countries
      description: Fetch all countries
      operationId: getCountries
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                description: Fetch all countries result
                required:
                  - content
                properties:
                  content:
                    description: Countries list
                    type: array
                    items:
                      $ref: '#/components/schemas/LocationViewDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '500':
          description: Internal server error
  /countries-count:
    get:
      tags:
        - Locations
      summary: Fetch list of all countries paired with their number of yachts
      description: Fetch list of all countries paired with their number of yachts
      operationId: getCountriesCount
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                description: Fetch all countries yacht counts result
                required:
                  - content
                properties:
                  content:
                    description: Countries and yachts counts list
                    type: array
                    items:
                      $ref: '#/components/schemas/CountryCountDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: 'boat4you_ws_common.openapi.yaml#/components/schemas/ErrorSchema'
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    LocationViewDto:
      type: object
      properties:
        id:
          type: string
          example: c-55
        realId:
          type: integer
          format: int64
          example: 55
        name:
          type: string
          example: Croatia
        locationType:
          type: string
          example: COUNTRY
        countryCode:
          type: string
          example: HR
...
    CountryCountDto:
      type: object
      properties:
        countryCode:
          type: string
          example: HR
        yachtCount:
          type: integer
          example: 25
...
security:
  - bearerAuth: []